requirements.txt
pymongo==3.12.3
djongo
django
dnspython
pytz
djangorestframework
pydantic